tools: fix persistent undo via desktop-mode
This has been something that has bothered me for a long time. I would
like to save my undo-history across Emacs restarts. I am using
`desktop-mode` which presumably should solve this, but it turns out that
the variables are fetching it's information from the function
`buffer-local-variables`. This function currently have a bug which omits
the data related to buffer-undo-list.

I created this advice in order to patch the output for the
`desktop-buffer-info` which is used to store the information about a
buffer so it can be restored later.